requirements.txt
Flask==3.1.2
flask-cors==6.0.2
python-dotenv==1.2.1

# Database
supabase==2.27.0
postgrest==2.27.0

# Blockchain
web3==7.14.0
eth-account==0.13.7
eth-utils==5.3.1

# HTTP
requests==2.32.5

# Data & ML (for predict.py)
numpy==2.2.6
pandas==2.3.3
scikit-learn==1.6.1
joblib==1.5.3
